Fix CA1034
